backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
attrs==25.3.0
bcrypt==4.3.0
boto3==1.40.30
botocore==1.40.30
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
fire==0.7.1
fonttools==4.59.2
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jmespath==1.0.1
lxml==6.0.1
motor==3.7.1
multidict==6.6.4
numpy==2.2.6
opencv-python-headless==4.12.0.88
passlib==1.7.4
pdf2docx==0.5.8
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.8
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymongo==4.15.0
PyMuPDF==1.26.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
reportlab==4.4.3
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
termcolor==3.1.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
pdf2image==1.17.0
pypandoc==1.14
pypandoc_binary==1.14
openai==1.58.1
